# app/docs/swagger.yaml
openapi: 3.0.0
info:
  title: Fisher Fans API
  description: API pour gérer les utilisateurs, bateaux, sorties de pêche, réservations et carnets de pêche.
  version: 1.0.0
servers:
  - url: https://api.fisherfans.com/
    description: Serveur de production Fisher Fans
  - url: https://localhost:3000/
    description: Serveur de développement Fisher Fans
paths:
...
  /v1/boats:
    post:
      summary: Créer un nouveau bateau
      tags:
        - Boats
      requestBody:
        $ref: '#/components/requestBodies/Boat'
      responses:
        '200':
          $ref: '#/components/responses/SuccessResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
    get:  # Nouveau endpoint pour récupérer des bateaux avec filtres
      summary: Obtenir une liste de bateaux avec des filtres
      tags:
        - Boats
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [open, cabin, catamaran, yacht, jetski, canoe]
          description: Filtrer par type de bateau
        - name: marque
          in: query
          schema:
            type: string
          description: Filtrer par marque
        - name: annee_fabrication
          in: query
          schema:
            type: integer
          description: Filtrer par année de fabrication
      responses:
        '200':
          description: Liste de bateaux
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Boat'
  /v1/boats/{id}:
    get:
      summary: Obtenir les détails d'un bateau spécifique
      tags:
        - Boats
      parameters:
        - $ref: '#/components/parameters/BoatId'
      responses:
        '200':
          $ref: '#/components/schemas/Boat'
        '404':
          $ref: '#/components/responses/ErrorResponse'
    put:
      summary: Modifier les informations d'un bateau
      tags:
        - Boats
      parameters:
        - $ref: '#/components/parameters/BoatId'
      requestBody:
        $ref: '#/components/requestBodies/Boat'
      responses:
        '200':
          $ref: '#/components/responses/SuccessResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
    delete:
      summary: Supprimer un bateau
      tags:
        - Boats
      parameters:
        - $ref: '#/components/parameters/BoatId'
      responses:
        '200':
          $ref: '#/components/responses/SuccessResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
...
  /v1/logs:
    post:
      summary: Créer une nouvelle page de carnet de pêche
      tags:
        - Logs
      requestBody:
        $ref: '#/components/requestBodies/Log'
      responses:
        '200':
          $ref: '#/components/responses/SuccessResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
  /v1/logs/{id}:
    get:
      summary: Obtenir les détails d'une page spécifique du carnet de pêche
      tags:
        - Logs
      parameters:
        - $ref: '#/components/parameters/LogId'
      responses:
        '200':
          $ref: '#/components/schemas/Log'
        '404':
          $ref: '#/components/responses/ErrorResponse'
    put:
      summary: Modifier une page du carnet de pêche
      tags:
        - Logs
      parameters:
        - $ref: '#/components/parameters/LogId'
      requestBody:
        $ref: '#/components/requestBodies/Log'
      responses:
        '200':
          $ref: '#/components/responses/SuccessResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
    delete:
      summary: Supprimer une page du carnet de pêche
      tags:
        - Logs
      parameters:
        - $ref: '#/components/parameters/LogId'
      responses:
        '200':
          $ref: '#/components/responses/SuccessResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
components:
  responses:
    SuccessResponse:
      description: Opération réussie
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/SuccessResponse'
    ErrorResponse:
      description: Requête incorrecte ou ressource non trouvée
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  requestBodies:
...
    Boat:
      description: Détails du bateau
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Boat'
...
    Log:
      description: Détails du carnet de pêche
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Log'
  parameters:
...
    BoatId:
      name: id
      in: path
      required: true
      schema:
        type: integer
      description: ID unique du bateau
...
    LogId:
      name: id
      in: path
      required: true
      schema:
        type: integer
      description: ID unique de la page du carnet de pêche
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          description: ID unique de l'utilisateur
        name:
          type: string
          description: Nom de l'utilisateur
        email:
          type: string
          description: Adresse e-mail de l'utilisateur
        role:
          type: string
          description: Rôle de l'utilisateur (particulier, professionnel)
          enum: [particulier, professionnel]
        boats:
          type: array
          items:
            $ref: '#/components/schemas/Boat'
    Boat:
      type: object
      properties:
        id:
          type: integer
          description: ID unique du bateau
        name:
          type: string
          description: Nom du bateau
        boat_type:
          type: string
          description: Type du bateau
          enum: [open, cabin, catamaran, yacht, jetski, canoe]
        motor:
          type: string
          description: Type de motorisation
          enum: [diesel, essence, none]
        capacity:
          type: integer
          description: Capacité du bateau
        owner:
          $ref: '#/components/schemas/User'
...
    Log:
      type: object
      properties:
        id:
          type: integer
          description: ID unique de la page du carnet de pêche
        fish_name:
          type: string
          description: Nom du poisson
        size:
          type: number
          format: float
          description: Taille du poisson en cm
        weight:
          type: number
          format: float
          description: Poids du poisson en kg
        location:
          type: string
          description: Lieu de la pêche
        date_caught:
          type: string
          format: date
          description: Date à laquelle le poisson a été pêché
        released:
          type: boolean
          description: Indique si le poisson a été relâché
        user:
          $ref: '#/components/schemas/User'
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          description: Indicateur de succès de l'opération
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Message d'erreur
...